Add attribute to set lock file type
An edge case for post-processing lock files to add attributes has us wanting to load not from .lock.json files but from .load.json files.  These files contain an attribute that we add to transparently set the policy group name.  We chose a different name because we can't in conscience change a 'lock' file even though the policyfile loader only uses these lock files to determine the name and revision of the archive to be loaded.